@@ -3,6 +3,12 @@
     "description": "A tool to automatically fix PHP code style",
     "license": "MIT",
     "type": "application",
+    "keywords": [
+        "fixer",
+        "standards",
+        "static analysis",
+        "static code analysis"
+    ],
     "authors": [
         {
             "name": "Fabien Potencier",
@@ -17,35 +23,36 @@
         "php": "^7.4 || ^8.0",
         "ext-json": "*",
         "ext-tokenizer": "*",
-        "composer/semver": "^3.2",
+        "composer/semver": "^3.3",
         "composer/xdebug-handler": "^3.0.3",
-        "doctrine/annotations": "^1.13",
-        "php-cs-fixer/diff": "^2.0",
+        "doctrine/annotations": "^2",
+        "doctrine/lexer": "^2 || ^3",
+        "sebastian/diff": "^4.0 || ^5.0",
         "symfony/console": "^5.4 || ^6.0",
         "symfony/event-dispatcher": "^5.4 || ^6.0",
         "symfony/filesystem": "^5.4 || ^6.0",
         "symfony/finder": "^5.4 || ^6.0",
         "symfony/options-resolver": "^5.4 || ^6.0",
-        "symfony/polyfill-mbstring": "^1.23",
-        "symfony/polyfill-php80": "^1.25",
-        "symfony/polyfill-php81": "^1.25",
+        "symfony/polyfill-mbstring": "^1.27",
+        "symfony/polyfill-php80": "^1.27",
+        "symfony/polyfill-php81": "^1.27",
         "symfony/process": "^5.4 || ^6.0",
         "symfony/stopwatch": "^5.4 || ^6.0"
     },
     "require-dev": {
         "justinrainbow/json-schema": "^5.2",
-        "keradus/cli-executor": "^1.5",
-        "mikey179/vfsstream": "^1.6.10",
-        "php-coveralls/php-coveralls": "^2.5.2",
+        "keradus/cli-executor": "^2.0",
+        "mikey179/vfsstream": "^1.6.11",
+        "php-coveralls/php-coveralls": "^2.5.3",
         "php-cs-fixer/accessible-object": "^1.1",
         "php-cs-fixer/phpunit-constraint-isidenticalstring": "^1.2",
         "php-cs-fixer/phpunit-constraint-xmlmatchesxsd": "^1.2.1",
-        "phpspec/prophecy": "^1.15",
+        "phpspec/prophecy": "^1.16",
         "phpspec/prophecy-phpunit": "^2.0",
         "phpunit/phpunit": "^9.5",
-        "phpunitgoodpractices/polyfill": "^1.5",
-        "phpunitgoodpractices/traits": "^1.9.1",
-        "symfony/phpunit-bridge": "^6.0",
+        "phpunitgoodpractices/polyfill": "^1.6",
+        "phpunitgoodpractices/traits": "^1.9.2",
+        "symfony/phpunit-bridge": "^6.2.3",
         "symfony/yaml": "^5.4 || ^6.0"
     },
     "suggest": {
@@ -66,6 +73,9 @@
         "php-cs-fixer"
     ],
     "config": {
+        "allow-plugins": {
+            "ergebnis/composer-normalize": true
+        },
         "sort-packages": true
     }
 }
