@@ -16,7 +16,8 @@
         }
     ],
     "require": {
-        "php": ">=8.0.2",
+        "php": ">=8.1",
+        "symfony/deprecation-contracts": "^2.1|^3",
         "symfony/polyfill-mbstring": "~1.0",
         "symfony/service-contracts": "^1.1|^2|^3",
         "symfony/string": "^5.4|^6.0"
