@@ -16,7 +16,10 @@
         }
     ],
     "require": {
-        "php": ">=8.0.2"
+        "php": ">=8.1"
+    },
+    "require-dev": {
+        "symfony/filesystem": "^6.0"
     },
     "autoload": {
         "psr-4": { "Symfony\\Component\\Finder\\": "" },
