@@ -3,9 +3,11 @@
 namespace App\Card;
 
 /**
- * Class Game21 represents the car game 21.
+ * Class Game21 represents the card game 21.
  *
  * @author Marie Grahn, grmstud@student.bth.se
+ *
+ * @SuppressWarnings(PHPMD.ExcessiveClassComplexity)
  */
 class Game21
 {
@@ -13,18 +15,38 @@ class Game21
     public const MAX_CARDS = 3;
 
     /**
-     * @var Deck   the deck of cards
-     * @var Dealer the dealer
-     * @var array  the array with Player21 objects
-     * @var int    the number of players, default 1, max 4
+     * @var Deck $deck the hand of the cards
+     */
+    private $deck;
+    /**
+     * @var Dealer $dealer the hand of the cards
+     */
+    private $dealer;
+    /**
+     * @var array<int, Player21> $players the hand of the cards
+     */
+    private $players;
+
+    /**
+     * * @var int    the number of players, default 1, max 4
+     */
+    private int $noOfPlayers;
+
+    /**
      * @var int    the number of cards to draw per round,
      *             default 1, max 3
      */
-    private $deck;
-    private $dealer;
-    private $players;
-    private $noOfPlayers;
-    private $noOfCards;
+    private int $noOfCards;
+
+    /**
+     * * @var bool    true if the game is over, false otherwise
+     */
+    private bool $gameover;
+
+    /**
+     * * @var string    the name of the winner
+     */
+    private string $winner;
 
     /**
      * Constructor to create an object of Game 21 with one dealer, one player,
@@ -37,6 +59,8 @@ class Game21
         $this->players = [new Player21('Spelare 1')];
         $this->noOfPlayers = 1;
         $this->noOfCards = 1;
+        $this->gameover = false;
+        $this->winner = "";
     }
 
     /**
@@ -48,11 +72,31 @@ class Game21
     }
 
     /**
+     * Get the number of players in the game.
+     *
+     * @return int
+     */
+    public function getNoOfPlayers(): int
+    {
+        return $this->noOfPlayers;
+    }
+
+    /**
+     * Get the deck of the cards.
+     *
+     * @return int
+     */
+    public function getNoOfCards(): int
+    {
+        return $this->noOfCards;
+    }
+
+    /**
      * Get the deck of the cards.
      *
      * @return Deck
      */
-    public function getDeck()
+    public function getDeck(): Deck
     {
         return $this->deck;
     }
@@ -62,7 +106,7 @@ class Game21
      *
      * @return Dealer
      */
-    public function getDealer()
+    public function getDealer(): Dealer
     {
         return $this->dealer;
     }
@@ -70,7 +114,7 @@ class Game21
     /**
      * Get the players.
      *
-     * @return array
+     * @return array<int, Player21>
      */
     public function getPlayers()
     {
@@ -78,16 +122,41 @@ class Game21
     }
 
     /**
+     * Get true if the game is over, false otherwise.
+     *
+     * @return bool
+     */
+    public function isItGameover(): bool
+    {
+        $gameover = false;
+        if ($this->checkIfAllAreContent()) {
+            $gameover = true;
+        }
+
+        return $gameover;
+    }
+
+    /**
+     * Get the name of the winner.
+     *
+     * @return string
+     */
+    public function getWinner(): string
+    {
+        return $this->winner;
+    }
+
+    /**
      * Initiate the game, shuffle the deck of cards. Set the number of players and
      * the number of cards to draw per round.
      *
-     * @var int the number of players, default 1, max 4
-     * @var int the number of cards to draw per round,
+     * @param int $players the number of players, default 1, max 4
+     * @param int $cards the number of cards to draw per round,
      *          default 1, max 3
      *
      * @return void
      */
-    public function initGame(int $players = 1, int $cards = 1)
+    public function initGame(int $players=1, int $cards=1): void
     {
         $this->deck->shuffle();
         $this->noOfPlayers = $players > self::MAX_PLAYERS ? self::MAX_PLAYERS : $players;
@@ -99,47 +168,58 @@ class Game21
 
     /**
      * Play the game.
+     * @SuppressWarnings(PHPMD.ElseExpression)
      *
      * @return string the result of the game, if any
      */
-    public function play()
+    public function play(): string
     {
         $result = '';
 
         // Play the game
         // Let the players play first
-        $this->playPlayers();
-        // Let the dealer play
-        $this->playDealer();
+        if (!$this->checkIfPlayersAreContent()) {
+            $this->playPlayers();
+        }
+
         // Calculate the result and return it
-        if ($this->checkIfAllAreContent()) {
-            $result = $this->result();
+        if ($this->checkIfPlayersAreContent()) {
+            // Let the dealer play when all the player are content
+            $this->playDealer();
+            if ($this->checkIfAllAreContent()) {
+                $result = $this->result();
+            }
         }
 
         return $result;
     }
 
-    private function playPlayers()
+    private function playPlayers(): void
     {
         for ($i = 0; $i < $this->noOfCards; ++$i) {
             foreach ($this->players as $player) {
                 if ('Nytt kort?' == $player->getResult() and !$player->isContent()) {
                     $card = $this->deck->getTopCard();
-                    $player->increaseHand($card);
+                    if ($card) {
+                        $player->increaseHand($card);
+                    }
                 }
             }
         }
         foreach ($this->players as $player) {
             $player->getSumOfHand();
+            $player->getResult();
         }
     }
 
-    private function playDealer()
+    private function playDealer(): void
     {
         for ($i = 0; $i < $this->noOfCards; ++$i) {
             if ('' == $this->dealer->getResult()) {
                 $card = $this->deck->getTopCard();
-                $this->dealer->increaseHand($card);
+                if ($card) {
+                    $this->dealer->increaseHand($card);
+                }
             }
             if ('' !== $this->dealer->getResult()) {
                 $this->dealer->setContent();
@@ -147,14 +227,42 @@ class Game21
         }
         $this->dealer->getSumOfHand();
     }
+    /**
+     * @SuppressWarnings(PHPMD.BooleanArgumentFlag)
+     */
+    public function getAllPlayersInfo(bool $jsonCards=false)
+    {
+        $allPlayersInfo = array();
+        // Add info about the dealer
+        $playerInfo = array();
+        $playerInfo['name'] = $this->dealer->getName();
+        $playerInfo['cards'] = $this->dealer->getHand()->getCards();
+        if ($jsonCards) {
+            $playerInfo['cards'] = $this->dealer->getHand()->getCardsJson();
+        }
+        $playerInfo['sum low/high'] = $this->dealer->getScoreLow() . '/' . $this->dealer->getScoreHigh();
+        $playerInfo['result'] = $this->dealer->getResult();
+        array_push($allPlayersInfo, $playerInfo);
+        // Add info about all the players
+        foreach ($this->players as $player) {
+            $playerInfo = array();
+            $playerInfo['name'] = $player->getName();
+            $playerInfo['cards'] = $player->getHand()->getCards();
+            if ($jsonCards) {
+                $playerInfo['cards'] = $player->getHand()->getCardsJson();
+            }
+            $playerInfo['sum low/high'] = $player->getScoreLow() . '/' . $player->getScoreHigh();
+            $playerInfo['result'] = $player->getResult();
+            array_push($allPlayersInfo, $playerInfo);
+        }
 
-    private function checkIfAllAreContent()
+        return $allPlayersInfo;
+    }
+
+    private function checkIfAllAreContent(): bool
     {
         $noOfContent = 0;
-        $noOfPlayers = count($this->players);
 
-        for ($i = 0; $i < $noOfPlayers; ++$i) {
-        }
         foreach ($this->players as $player) {
             if ($player->isContent()) {
                 ++$noOfContent;
@@ -166,30 +274,47 @@ class Game21
 
         $total = count($this->players) + 1; // dealer included
 
-        return $total == $noOfContent ? true : false;
+        return $total === $noOfContent ? true : false;
     }
 
-    public function result()
+    private function checkIfPlayersAreContent(): bool
+    {
+        $noOfContent = 0;
+
+        foreach ($this->players as $player) {
+            if ($player->isContent()) {
+                ++$noOfContent;
+            }
+        }
+        $total = count($this->players);
+
+        return $total === $noOfContent ? true : false;
+    }
+
+    /**
+     * @SuppressWarnings(PHPMD.ElseExpression)
+     */
+    public function result(): string
     {
         $result = 'Vinnaren är: ';
         $winner = $this->dealer->getName();
         $bestScore = $this->dealer->getBestScore();
-        $noOfCards = $this->dealer->getNoOfCards();
-        $noOfPlayers = count($this->players);
-        for ($i = 0; $i < $noOfPlayers; ++$i) {
-            $playerBestScore = $this->players[$i]->getBestScore();
+        foreach ($this->players as $player) {
+            $playerBestScore = $player->getBestScore();
+            $noOfCardsBestScore = $player->getNoOfCards();
             if ($playerBestScore <= 21 and $playerBestScore > $bestScore) {
-                if (($playerBestScore == $bestScore) and ($noOfCards > $this->players[$i]->getNoOfCards())) {
-                    $winner = $this->players[$i]->getName();
+                $winner = $player->getName();
+                $bestScore = $playerBestScore;
+                if ($playerBestScore === $bestScore and $noOfCardsBestScore > $player->getNoOfCards()) {
+                    $winner = $player->getName();
                     $bestScore = $playerBestScore;
-                } elseif ($playerBestScore == $bestScore and $noOfCards == $this->players[$i]->getNoOfCards()) {
-                    $winner = $winner . ' & ' . $this->players[$i]->getName();
-                } else {
-                    $winner = $this->players[$i]->getName();
-                    $bestScore = $playerBestScore;
+                } elseif ($playerBestScore === $bestScore and $noOfCardsBestScore === $player->getNoOfCards()
+                    and ! str_contains($winner, $player->getName())) {
+                    $winner = $winner . ' & ' . $player->getName();
                 }
             }
         }
+        $this->winner = $winner;
 
         return $result . '<br>' . $winner;
     }
