@@ -13,12 +13,15 @@ class Game21Test extends KernelTestCase
      * Construct object and verify that the object is of expected instance.
      * Use no argument.
      */
-    public function testCreateObjectWithArgument()
+    public function testCreateObject()
     {
         $game = new Game21();
         $this->assertInstanceOf("\App\Card\Game21", $game);
-        $this->assertEquals($game->getDealer()->getName(), 'Banken');
-        $this->assertEquals($game->getPlayers()[0]->getName(), 'Spelare 1');
+        $this->assertEquals('Banken', $game->getDealer()->getName());
+        $this->assertEquals('Spelare 1', $game->getPlayers()[0]->getName());
+        $this->assertEquals(1, $game->getNoOfPlayers());
+        $this->assertEquals(1, $game->getNoOfCards());
+        $this->assertEquals(52, $game->getDeck()->getNoOfCards());
     }
 
     /**
@@ -34,92 +37,14 @@ class Game21Test extends KernelTestCase
     }
 
     /**
-     * Construct object and test to play the game with 1 player. Both player and
-     * bank has 21 => bank is the winner.
+     * Construct object and get all json information.
      */
-    public function testPlayEqualBankWins()
+    public function testCreateGameAndGetJsonInfo()
     {
         $game = new Game21();
-        $result = $game->play();
-        $this->assertEquals($game->getDealer()->getName(), 'Banken');
-        $this->assertEquals($game->getPlayers()[0]->getName(), 'Spelare 1');
-        $game->getDealer()->setContent();
-        $game->getDealer()->setScoreLow(21);
-        $game->getPlayers()[0]->setContent();
-        $game->getPlayers()[0]->setScoreLow(21);
-        $result = $game->play();
-        $this->assertEquals($result, 'Vinnaren är: <br>Banken');
-    }
-
-    /**
-     * Construct object and test to play the game with 1 player. Both player has
-     * 20 and bank has 21 => bank is the winner.
-     */
-    public function testPlayBankWins()
-    {
-        $game = new Game21();
-        $result = $game->play();
-        $this->assertEquals($game->getDealer()->getName(), 'Banken');
-        $this->assertEquals($game->getPlayers()[0]->getName(), 'Spelare 1');
-        $game->getDealer()->setContent();
-        $game->getDealer()->setScoreLow(21);
-        $game->getPlayers()[0]->setContent();
-        $game->getPlayers()[0]->setScoreLow(20);
-        $result = $game->play();
-        $this->assertEquals($result, 'Vinnaren är: <br>Banken');
-    }
-
-    /**
-     * Construct object and test to play the game with 1 player. Both player has
-     * 20 and bank has 18 => player is the winner.
-     */
-    public function testPlayPlayerWins()
-    {
-        $game = new Game21();
-        $result = $game->play();
-        $this->assertEquals($game->getDealer()->getName(), 'Banken');
-        $this->assertEquals($game->getPlayers()[0]->getName(), 'Spelare 1');
-        $game->getDealer()->setContent();
-        $game->getDealer()->setScoreLow(18);
-        $game->getPlayers()[0]->setContent();
-        $game->getPlayers()[0]->setScoreLow(20);
-        $result = $game->play();
-        $this->assertEquals($result, 'Vinnaren är: <br>Spelare 1');
-    }
-
-    /**
-     * Construct object and test to play the game with 1 player. Both player has
-     * 20 and bank has more than 21 => player is the winner.
-     */
-    public function testPlayBankThickPlayerWins()
-    {
-        $game = new Game21();
-        $result = $game->play();
-        $this->assertEquals($game->getDealer()->getName(), 'Banken');
-        $this->assertEquals($game->getPlayers()[0]->getName(), 'Spelare 1');
-        $game->getDealer()->setContent();
-        $game->getDealer()->setScoreLow(22);
-        $game->getPlayers()[0]->setContent();
-        $game->getPlayers()[0]->setScoreLow(20);
-        $result = $game->play();
-        $this->assertEquals($result, 'Vinnaren är: <br>Spelare 1');
-    }
-
-    /**
-     * Construct object and test to play the game with 1 player. Both player has
-     * 22 and bank has  20 => bank is the winner.
-     */
-    public function testPlayPlayerThickBankWins()
-    {
-        $game = new Game21();
-        $result = $game->play();
-        $this->assertEquals($game->getDealer()->getName(), 'Banken');
-        $this->assertEquals($game->getPlayers()[0]->getName(), 'Spelare 1');
-        $game->getDealer()->setContent();
-        $game->getDealer()->setScoreLow(20);
-        $game->getPlayers()[0]->setContent();
-        $game->getPlayers()[0]->setScoreLow(22);
-        $result = $game->play();
-        $this->assertEquals($result, 'Vinnaren är: <br>Banken');
+        $this->assertEquals("Banken", $game->getAllPlayersInfo(true)[0]["name"]);
+        $this->assertEquals("", $game->getAllPlayersInfo(true)[0]["result"]);
+        $this->assertEquals("Spelare 1", $game->getAllPlayersInfo(true)[1]["name"]);
+        $this->assertEquals("Nytt kort?", $game->getAllPlayersInfo(true)[1]["result"]);
     }
 }
